service-workers/service-worker/skip-waiting-without-using-registration.https.html

<!DOCTYPE html>
<title>Service Worker: Skip waiting without using registration</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

promise_test(async t => {
  const SCOPE = 'resources/blank.html?skip-waiting-without-using-registration';
  const SCRIPE = 'resources/skip-waiting-worker.js';
  await service_worker_unregister(t, SCOPE);
  const frame = await with_iframe(SCOPE);
  const serviceWorker = await frame.contentWindow.navigator.serviceWorker;
  assert_equals(serviceWorker.controller, null,
                'Document controller should be null');
  const registration =
      await navigator.serviceWorker.register(SCRIPE, {scope: SCOPE});
  await wait_for_state(t, registration.installing, 'activated');

  t.add_cleanup(async() => {
    if (registration)
      await registration.unregister();
    if (frame)
      frame.remove();
  });

  assert_equals(serviceWorker.controller, null,
                'Document controller should still be null');
  assert_not_equals(registration.active, null,
                    'Registration active worker should not be null');
  fetch_tests_from_worker(registration.active);
}, 'Test skipWaiting while a client is not being controlled');

</script>
